requirements.txt
-e .

av
awscli
black
cloudpathlib[s3]
decord
fastai
ffmpeg-python
flake8
future
fvcore
gitpython
ipython
jedi<0.18.0
jupyter
jupyterlab>=3.1.4
loguru
matplotlib
nb_black
nbautoexport
numpy
onnx
onnxruntime-gpu
opencv-python-headless
openpyxl
pandas>1.2.0
pandas_path
pip
progressbar
pydantic
pytest
python-dotenv
pytorch-lightning
pytorchvideo
scikit-learn
seaborn
setuptools
tensorboard
timm
torchinfo
torchvision>=0.10.0
tqdm
typer[all]
wheel
